#267: Renamed the file containing the Caracteristiques
... to follow the conventions.
