+++ src/app/ffbe/mappers/consumable-mapper.ts
@@ -1,6 +1,6 @@
 import {Consumable} from '../model/consumable.model';
 import {Objet} from '../model/objet/objet.model';
-import {Caracteristiques} from '../model/caracteristiques';
+import {Caracteristiques} from '../model/caracteristiques.model';
 import {ObjetElements} from '../model/objet/objet-elements';
 import {FFBE_ENGLISH_TABLE_INDEX, FFBE_FRENCH_TABLE_INDEX} from '../ffbe.constants';
 import {ItemCategoryFactory} from '../model/item-category.model';
