+++ src/app/ffbe/model/caracteristiques.model.ts
